post for specific User in admin and more security

+++ admin/includes/view_all_users.php
@@ -62,14 +62,19 @@ if (isset($_GET['change_to_sub'])) {
 }
 
 if (isset($_GET['delete'])) {
-    $the_user_id = $_GET['delete'];
+    if (isset($_SESSION['user_role'])){
+        if ($_SESSION['user_role'] === 'admin'){
+            $the_user_id = mysqli_real_escape_string($connection,$_GET['delete']) ;
 
-    $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
-    $delete_query = mysqli_query($connection, $query);
+            $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
+            $delete_query = mysqli_query($connection, $query);
+        
+            confirmQuery($delete_query);
+        
+            header("Location: users.php");
+        }
+    }
 
-    confirmQuery($delete_query);
-
-    header("Location: users.php");
 }
 
 ?>
